requirements.txt
# ====================================
# S.P.E.C. Valuation Engine - Dependencies
# ====================================

# Core Framework
streamlit>=1.29.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
pyarrow>=14.0.0  # For Parquet support

# Machine Learning (Pinned for SHAP compatibility)
xgboost>=1.7.0,<2.0.0  # Must be 1.x for SHAP 0.42.x compatibility
scikit-learn>=1.3.0
llvmlite==0.42.0  # Pre-built wheel
numba==0.59.1     # Pre-built wheel
shap==0.42.1      # Pinned for pre-built wheel compatibility

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0

# AI/LLM Integration
openai>=1.6.0

# Utilities
python-dotenv>=1.0.0

# Code Quality (Dev)
black>=23.0.0
ruff>=0.1.0
